requirements.txt
aiohttp==3.8.1
aiogram==2.20
beautifulsoup4==4.10.0
pycryptodome==3.14.1
